@@ -7,8 +7,8 @@
     "test": "test"
   },
   "dependencies": {
-    "@quantadex/bitsharesjs": "^1.8.3",
-    "@quantadex/bitsharesjs-ws": "^1.5.5",
+    "@quantadex/bitsharesjs": "^1.8.8",
+    "@quantadex/bitsharesjs-ws": "^1.5.6",
     "@quantadex/quanta-base": "^0.8.3",
     "@quantadex/quanta_js": "^1.0.6",
     "autosuggest-highlight": "^3.1.1",
@@ -19,58 +19,67 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "backpack-core": "^0.5.0",
-    "bignumber.js": "^8.0.1",
+    "bignumber.js": "^8.1.1",
+    "bs58": "^4.0.1",
     "counterpart": "^0.18.6",
     "emotion": "^8.0.12",
     "express": "^4.16.4",
-    "highcharts": "^7.0.2",
+    "highcharts": "^7.1.1",
     "js-sorted-set": "^0.2.0",
-    "jspdf": "^1.4.1",
+    "jspdf": "^1.5.3",
     "mobx": "^3.6.1",
     "mobx-react": "^4.4.3",
-    "mobx-react-form": "^1.32.3",
-    "moment": "^2.22.1",
+    "mobx-react-form": "^1.39.1",
+    "moment": "^2.24.0",
     "numeral": "^2.0.6",
-    "react": "^16.6.3",
-    "react-autosuggest": "^9.3.4",
-    "react-dom": "^16.6.3",
+    "qrcode": "^1.3.3",
+    "react": "^16.8.6",
+    "react-autosuggest": "^9.4.3",
+    "react-bootstrap": "^1.0.0-beta.8",
+    "react-dom": "^16.8.6",
     "react-emotion": "^9.1.3",
-    "react-ga": "^2.5.6",
+    "react-ga": "^2.5.7",
+    "react-google-invisible-recaptcha": "^0.2.11",
     "react-numeric-input": "^2.2.3",
     "react-quill": "^1.2.7",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
-    "react-toastify": "^4.4.3",
+    "react-slider": "^0.11.2",
+    "react-toastify": "^4.5.2",
+    "react-tooltip": "^3.10.0",
     "redux": "^3.7.2",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
+    "rollbar": "^2.6.1",
     "stellar-base": "0.9.0",
     "stellar-sdk": "^0.8.2",
     "validatorjs": "^3.14.2",
+    "wallet-address-validator": "^0.2.4",
+    "web3": "^1.0.0-beta.52",
     "webpack": "^3.12.0",
     "webpack-cli": "^2.1.5",
-    "webpack-dev-server": "^3.1.10",
-    "whatwg-fetch": "^2.0.3",
+    "webpack-dev-server": "^3.3.1",
+    "whatwg-fetch": "^2.0.4",
     "with-query": "^1.0.2",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@webpack-cli/migrate": "^0.1.2",
+    "@webpack-cli/migrate": "^0.1.5",
     "babel-core": "^6.26.3",
     "babel-loader": "^7.1.5",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
-    "css-loader": "^2.0.1",
+    "css-loader": "^2.1.1",
     "html-webpack-plugin": "^3.2.0",
     "json-loader": "^0.5.7",
-    "redux-devtools": "^3.4.1",
+    "redux-devtools": "^3.5.0",
     "redux-devtools-dock-monitor": "^1.1.3",
     "redux-devtools-log-monitor": "^1.4.0",
     "style-loader": "^0.23.1",
     "uglifyjs-webpack-plugin": "^1.2.3",
-    "webpack-merge": "^4.1.2"
+    "webpack-merge": "^4.2.1"
   },
   "scripts": {
     "build:prod": "node_modules/.bin/webpack --config webpack.prod.js",
