@@ -108,7 +108,7 @@ const ApplicationApi = {
                 ] = res;
 
                 let memo_from_public, memo_to_public;
-                if (memo && encrypt_memo) {
+                if (memo) {
                     memo_from_public = chain_memo_sender.getIn([
                         "options",
                         "memo_key"
@@ -168,7 +168,7 @@ const ApplicationApi = {
                               )
                             : Buffer.isBuffer(memo)
                                 ? memo.toString("utf-8")
-                                : memo
+                                : Buffer.from(memo)
                     };
                 }
                 // Allow user to choose asset with which to pay fees #356
@@ -195,7 +195,7 @@ const ApplicationApi = {
                     amount: {amount, asset_id: chain_asset.get("id")},
                     memo: memo_object
                 });
-
+                
                 return tr.update_head_block().then(() => {
                     if (propose_account) {
                         tr.add_type_operation("proposal_create", {
@@ -210,6 +210,230 @@ const ApplicationApi = {
                 });
             })
             .catch(() => {});
+    },
+
+    vesting_balance_withdraw({
+        // OBJECT: { ... }
+        account,
+        balance,
+        amount,
+        asset,
+        fee_asset_id = "1.3.0"
+    }) {
+        let unlock_promise = new Promise((resolve, reject) => resolve());
+
+        return Promise.all([
+            FetchChain("getAccount", account),
+            FetchChain("getObject", balance),
+            FetchChain("getAsset", asset),
+            FetchChain("getAsset", fee_asset_id),
+            unlock_promise
+        ])
+            .then(res => {
+                let [
+                    chain_owner,
+                    chain_balance,
+                    chain_asset,
+                    chain_fee_asset
+                ] = res;
+
+                // Allow user to choose asset with which to pay fees #356
+                let fee_asset = chain_fee_asset.toJS();
+
+                // Default to CORE in case of faulty core_exchange_rate
+                if (
+                    fee_asset.options.core_exchange_rate.base.asset_id ===
+                        "1.3.0" &&
+                    fee_asset.options.core_exchange_rate.quote.asset_id ===
+                        "1.3.0"
+                ) {
+                    fee_asset_id = "1.3.0";
+                }
+
+                // https://bitshares.org/doxygen/structgraphene_1_1chain_1_1balance__claim__operation.html
+                let tr = new TransactionBuilder();
+                let vesting_balance_withdraw_op = tr.get_type_operation("vesting_balance_withdraw", {
+                    fee: {
+                        amount: 0,
+                        asset_id: fee_asset_id
+                    },
+                    vesting_balance: chain_balance.get("id"),
+                    owner: chain_owner.get("id"),
+                    amount: {amount, asset_id: chain_asset.get("id")},
+                });
+                
+                return tr.update_head_block().then(() => {
+                    tr.add_operation(vesting_balance_withdraw_op);
+                    return tr;
+                });
+            })
+            .catch(() => {});
+    },
+
+    balance_claim({
+        // OBJECT: { ... }
+        account,
+        public_key,
+        balance,
+        amount,
+        asset,
+        fee_asset_id = "1.3.0"
+    }) {
+        let unlock_promise = new Promise((resolve, reject) => resolve());
+
+        return Promise.all([
+            FetchChain("getAccount", account),
+            FetchChain("getObject", balance),
+            FetchChain("getAsset", asset),
+            FetchChain("getAsset", fee_asset_id),
+            unlock_promise
+        ])
+            .then(res => {
+                let [
+                    chain_owner,
+                    chain_balance,
+                    chain_asset,
+                    chain_fee_asset
+                ] = res;
+
+                // Allow user to choose asset with which to pay fees #356
+                let fee_asset = chain_fee_asset.toJS();
+
+                // Default to CORE in case of faulty core_exchange_rate
+                if (
+                    fee_asset.options.core_exchange_rate.base.asset_id ===
+                        "1.3.0" &&
+                    fee_asset.options.core_exchange_rate.quote.asset_id ===
+                        "1.3.0"
+                ) {
+                    fee_asset_id = "1.3.0";
+                }
+
+                // https://bitshares.org/doxygen/structgraphene_1_1chain_1_1balance__claim__operation.html
+                let tr = new TransactionBuilder();
+                let balance_claim_op = tr.get_type_operation("balance_claim", {
+                    fee: {
+                        amount: 0,
+                        asset_id: fee_asset_id
+                    },
+                    balance_to_claim: chain_balance.get("id"),
+                    deposit_to_account: chain_owner.get("id"),
+                    balance_owner_key: public_key,
+                    total_claimed: {amount, asset_id: chain_asset.get("id")},
+                });
+                
+                return tr.update_head_block().then(() => {
+                    tr.add_operation(balance_claim_op);
+                    return tr;
+                });
+            })
+            .catch(() => {});
+    },
+
+    account_upgrade({
+        // OBJECT: { ... }
+        account,
+        fee_asset_id = "1.3.0"
+    }) {
+        let unlock_promise = new Promise((resolve, reject) => resolve());
+
+        return Promise.all([
+            FetchChain("getAccount", account),
+            FetchChain("getAsset", fee_asset_id),
+            unlock_promise
+        ])
+            .then(res => {
+                let [
+                    chain_account,
+                    chain_fee_asset
+                ] = res;
+
+                // Allow user to choose asset with which to pay fees #356
+                let fee_asset = chain_fee_asset.toJS();
+
+                // Default to CORE in case of faulty core_exchange_rate
+                if (
+                    fee_asset.options.core_exchange_rate.base.asset_id ===
+                        "1.3.0" &&
+                    fee_asset.options.core_exchange_rate.quote.asset_id ===
+                        "1.3.0"
+                ) {
+                    fee_asset_id = "1.3.0";
+                }
+
+                let tr = new TransactionBuilder();
+                let account_upgrade_op = tr.get_type_operation("account_upgrade", {
+                    fee: {
+                        amount: 0,
+                        asset_id: fee_asset_id
+                    },
+                    account_to_upgrade: chain_account.get("id"),
+                    upgrade_to_lifetime_member: true,
+                });
+                
+                return tr.update_head_block().then(() => {
+                    tr.add_operation(account_upgrade_op);
+                    return tr;
+                });
+            })
+            .catch(() => {});
+    },
+
+    roll_dice({
+        account,
+        amount,
+        asset,
+        bet,
+        numbers = [],
+        fee_asset_id = "1.3.0",
+    }) {
+        let unlock_promise = new Promise((resolve, reject) => resolve());
+        return Promise.all([
+            FetchChain("getAccount", account),
+            FetchChain("getAsset", asset),
+            FetchChain("getAsset", fee_asset_id),
+            unlock_promise
+        ])
+            .then(res => {
+                let [
+                    chain_account,
+                    chain_asset,
+                    chain_fee_asset
+                ] = res;
+
+                // Allow user to choose asset with which to pay fees #356
+                let fee_asset = chain_fee_asset.toJS();
+
+                // Default to CORE in case of faulty core_exchange_rate
+                if (
+                    fee_asset.options.core_exchange_rate.base.asset_id ===
+                        "1.3.0" &&
+                    fee_asset.options.core_exchange_rate.quote.asset_id ===
+                        "1.3.0"
+                ) {
+                    fee_asset_id = "1.3.0";
+                }
+
+                let tr = new TransactionBuilder();
+                let roll_dice_op = tr.get_type_operation("roll_dice", {
+                    account_id: chain_account.get("id"),
+                    fee: {
+                        amount: 100,
+                        asset_id: fee_asset_id
+                    },
+                    risk: {
+                        amount,
+                        asset_id: chain_asset.get("id")
+                    },
+                    bet,
+                    numbers
+                });
+                
+                return tr.update_head_block().then(() => {
+                    tr.add_operation(roll_dice_op);
+                    return tr;
+                });
+            })
     },
 
     issue_asset(
