@@ -1,23 +1,20 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux'
-import moment from 'moment';
-import { SmallToken } from './ui/ticker.jsx'
-
 import { css } from 'emotion'
-import globalcss from './global-css.js'
 import QTTabBar from './ui/tabBar.jsx'
-import QTTableViewSimple from './ui/tableViewSimple.jsx'
-import {switchTicker} from "../redux/actions/app.jsx";
+import {SymbolToken} from './ui/ticker.jsx'
 import SearchBox from "./ui/searchBox.jsx";
+import { withRouter} from "react-router-dom";
+import ReactTooltip from 'react-tooltip'
+import Loader from './ui/loader.jsx'
 
 const container = css`
 	padding: 20px;
 	flex: auto;
 	overflow: auto;
-	border-right: 1px solid #333;
-	.coin-tabbar {
-		padding:10px 21px;
-	}
+	border: 3px solid #000;
+	background-color: #23282c;
+	cursor: initial;
 
  	table {
 	  width: 100%;
@@ -25,7 +22,7 @@ const container = css`
 		  color: #777;
 	  }
 	  tr {
-		  border-bottom: 1px solid #222;
+		  border-bottom: 1px solid #333;
 		  cursor: pointer;
 	  }
 	  tbody tr:hover {
@@ -47,6 +44,9 @@ const container = css`
 	}
 	
 	&.mobile {
+		border: 0;
+		background: transparent;
+		
 		h4 {
 			display: none;
 		}
@@ -55,6 +55,7 @@ const container = css`
 		}
 		table tr {
 			border-bottom: 1px solid #333;
+			line-height: 40px;
 		}
 	}
 `;
@@ -65,92 +66,119 @@ class Dashboard extends Component {
 		super(props)
 		this.state = {
 			selectedTab:"All",
+			selectedCoin: 0,
 			filter: ""
 		}
 	}
 
 	switchMarket(e) {
-		this.props.dispatch(switchTicker(e))
-		if (this.props.mobile) {
-			const list = document.getElementById("market-list")
-			if (list.classList.contains("active")) {
-				list.classList.remove("active")
-			} else {
-				list.classList.add("active")
-			}
-		}
+		const { mobile_nav, closeSelf, history, match, location } = this.props
+
+		history.push("/" + (match.params.net || "mainnet") + "/exchange/" + e.replace("/", "_") + location.search)
+		setTimeout(() => {
+			if (mobile_nav) mobile_nav()
+			if (closeSelf) closeSelf()
+		}, 0)
 	}
 
 	handleChange(e) {
 		this.setState({filter: e.target.value})
 	}
 
-	render() {
-		const tabs = {
-			names: ['All','Trading','Favorites'],
-			selectedTabIndex: 0
-		}
+	switchCoin(index) {
+		this.setState({selectedCoin: index})
+	}
 
+	shortName(market) {
+		const pairs = market.split('/')
+		const base = pairs[0].split('0X')
+		const counter = pairs[1].split('0X')
+		
+		return `${base[0]}${base[1] ? "0X" + base[1].substr(0,4) : ""}/${counter[0]}${counter[1] ? "0X" + counter[1].substr(0,4) : ""}`
+	}
+
+	render() {
+		const { markets, mobile, announcements } = this.props
+		
 		const subtabs = {
-			names: ['BTC','ETH','XMR','USDT'],
+			names: Object.keys(markets),
 			selectedTabIndex: 0,
 		}
-
+		
 		return (
-			<div className={container + (this.props.mobile ? " mobile" : "")}>
-        {/* <section className="menu d-flex justify-content-start qt-font-extra-small qt-font-light text-center">
+			<div className={container + (mobile ? " mobile qt-font-small" : "")} onClick={e => e.stopPropagation()}>
+				<div className="d-flex justify-content-between border-bottom border-dark mb-3">
+					<h4>MARKETS</h4>
+					<SearchBox onChange={this.handleChange.bind(this)} placeholder="Search Pairs" 
+						style={{margin: "-7px 0 7px", border: 0, backgroundColor: "rgba(255,255,255,0.07)"}} />
+				</div>
+				{ announcements ?
+					<div className="mb-5" style={{margin: "0 -20px"}}>
+						{announcements}
+					</div>
+				: null
+				}
+				
+				{/* <section className="menu d-flex justify-content-start qt-font-extra-small qt-font-light text-center">
 					<QTTabBar
 						className="block medium fluid qt-font-regular d-flex justify-content-start"
-					 	tabs = {tabs}
-					/>
-        </section>
-				<section className="coin-tabbar">
-					<QTTabBar
-						className="underline small fluid even-width qt-font-bold d-flex justify-content-between"
-						width={58.3}
-						tabs = {subtabs}
+						tabs = {tabs}
 					/>
 				</section> */}
+				<section className="mb-3">
+					<QTTabBar
+						className="underline small fluid even-width qt-font-bold d-flex justify-content-start"
+						width={58.3}
+						tabs={subtabs}
+						switchTab={this.switchCoin.bind(this)}
+					/>
+				</section>
 				
-        <section className="price">
-			<div className="d-flex justify-content-between border-bottom border-dark mb-3">
-				<h4>MARKETS</h4>
-				<SearchBox onChange={this.handleChange.bind(this)} placeholder="Search Pairs" 
-					style={{margin: "-7px 0 7px", border: 0, backgroundColor: "rgba(255,255,255,0.07)"}} />
-			</div>
-			
-			<table>
-				<thead>
-					<tr>
-						<td>Pairs</td>
-						<td className="text-right">Price</td>
-						<td className="text-right">Volume</td>
-					</tr>
-				</thead>
-				<tbody>
-							{
-								this.props.markets.filter(market => market.name.toLowerCase().includes(this.state.filter)).map((market, index) => {
-									return <tr key={index} onClick={this.switchMarket.bind(this, market.name)}>
-										<td className="market">
-											{market.name}
-										</td>
-										<td className="text-right">{market.last}</td>
-										<td className="text-right">{market.base_volume}</td>
-									</tr>
-								})
-							}
-				</tbody>
-			</table>
-        </section>
+				{ markets.length != 0 ?
+					<section className="price">
+						
+						<table>
+							<thead>
+								<tr>
+									<td>Pairs</td>
+									<td className="text-right">Price</td>
+									<td className="text-right">
+										<img data-tip="The market depth for this<br/> token on the QUANTA Network"
+											src={devicePath("public/images/question.png")} /> Depth
+									</td>
+								</tr>
+							</thead>
+							<tbody>
+								{
+									markets[subtabs.names[this.state.selectedCoin]] && markets[subtabs.names[this.state.selectedCoin]].filter(market => this.shortName(market.name).toLowerCase().includes(this.state.filter.toLowerCase())).map((market, index) => {
+										let pair = market.name.split('/')
+										let usd_price = pair[1] === "ETH" ? window.binance_price["ETHUSDT"] 
+														: (pair[1] === "BTC" ? window.binance_price["BTCUSDT"] 
+															: pair[1].startsWith("TUSD") || pair[1] == "USD" ? 1 : 0)
+										return <tr key={index} onClick={() => this.switchMarket(market.name)}>
+											<td className="market">
+												<SymbolToken name={pair[0]} showIcon={false} withLink={false} />
+											</td>
+											<td className="text-right">{market.last == 0 ? "-" : market.last}</td>
+											<td className="text-right">{window.allMarketsByHash[market.name].depth == 0 || usd_price == 0 ? "-" : "$" + Math.round(window.allMarketsByHash[market.name].depth * parseFloat(usd_price))}</td>
+										</tr>
+									})
+								}
+							</tbody>
+						</table>
+					</section>
+				: <Loader size="50px" />
+				}
+				<ReactTooltip clickable={true} html={true} />
 			</div>
 		);
 	}
 }
 
+// lodash.sortBy(this.props.markets[subtabs.names[this.state.selectedCoin]], 'base_volume')
 
 const mapStateToProps = (state) => ({
-	markets: state.app.markets,
-	favoriteList: state.app.favoriteList
+	markets: state.app.markets
 });
 
-export default connect(mapStateToProps)(Dashboard);
+export default connect(mapStateToProps)(withRouter (Dashboard));
