@@ -1,89 +1,135 @@
 import React, { Component } from 'react';
 import Header from './header.jsx';
+import Announcement from './announcement.jsx'
 import Chart from './chart.jsx';
 import DepthChart from './chart_depth.jsx';
 import TradingHistory from './trading_history.jsx';
 import OrderBook from './order_book.jsx';
-import Dashboard from './dashboard.jsx';
 import Menu from './menu.jsx';
 import Orders from './orders.jsx';
 import Trade from './trade.jsx';
-import Leaderboard from './leaderboard.jsx'
+import Balance from './balance.jsx';
+import Connect from './connect.jsx';
 import Status from './status.jsx'
-import FirstTime from './first_time.jsx'
-import QTTableView from './ui/tableView.jsx'
-import Order from './order.jsx';
-import Markets from './markets.jsx';
-import OpenOrders from './open_orders.jsx';
-
-import {switchTicker, initBalance, getMarketQuotes} from "../redux/actions/app.jsx";
+import Switch from './ui/switch.jsx';
+import { switchTicker, reconnectIfNeeded } from '../redux/actions/app.jsx'
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import { connect } from 'react-redux'
-
 import { css } from 'emotion'
-import globalcss from './global-css.js'
-
-import { Link } from 'react-router-dom'
-
-import QTTableViewSimple from './ui/tableViewSimple.jsx'
+import CONFIG from '../config.js'
 
 const container = css`
-	background-color:${globalcss.COLOR_BACKGROUND};
+	background-color: #121517;
 	position: relative;
+	height: 100vh;
 
-	.exchange-left {
-		width:281px;
-		padding:22px 20px;
+	hr {
+		border-color: #444;
 	}
 
-	.exchange-middle {
-		position: relative;
-		padding: 21px 19px;
-		flex-grow:1;
-		background-color:rgba(17,20,22,1);
-		padding-bottom: 0;
-	}
-
-	.exchange-right {
-		width:300px;
-		background-color: rgba(0,0,0,0.26)
-	}
-
-	.exchange-bottom {
-		bottom: 0;
+	section.compartment {
 		background-color: #23282c;
+		margin-right: 5px;
+		min-height: 260px;
+	}
+
+	.content {
+		.cols, .left-cols {
+			flex: 0 0 260px;
+		}
+		.left-cols {
+			padding: 10px;
+			height: calc(100vh - 92px);
+		}
+	}
+
+	.status {
+		position: fixed;
+		bottom: 0;
 		justify-content: center;
-		border-top: 3px solid black;
+		width: 100%;
+		background-color: #23282c;
 		z-index: 99;
 	}
 	
 	#tv_chart_container, #depth_chart_container {
-		height: calc(100vh - 530px);
-		min-height: 370px !important;
-		padding-bottom: 20px;
-		border-bottom: 1px solid #333;
+		height: calc(100vh - 124px);
+		width: 100%;
+	}
+
+	&.has-user {
+		#tv_chart_container, #depth_chart_container {
+			height: calc(100vh - 390px);
+		}
+	}
+
+	&.has-announcement {
+		.left-cols {
+			height: calc(100vh - 120px);
+		}
+
+		#tv_chart_container, #depth_chart_container {
+			height: calc(100vh - 152px);
+		}
+
+		.orderbook-ask, .orderbook-bid {
+			height: calc(50vh - 110px);
+		}
+
+		&.has-user #tv_chart_container, &.has-user #depth_chart_container {
+			height: calc(100vh - 417px);
+		}
+	}
+
+	.trade-toggle {
+		position: absolute;
+		right: 10px;
+		float: right;
+		display: flex;
+		font-size: 16px;
+		padding: 8px 10px 0px
+		
+		.toggle {
+			margin-left: 73px;
+			color: #66d7d7;
+			padding-left: 14px;
+			font-size: 12px;
+		}
+		.toggle.show {
+			background: url(${devicePath('public/images/left-arrow.svg')}) no-repeat 0 4px;
+		}
+		.toggle.hide {
+			margin-left: 80px;
+			background: url(${devicePath('public/images/right-arrow.svg')}) no-repeat 0 4px;
+		}
 	}
 
 	.switch-chart {
-		position: absolute;
-		right: 20px;
-		margin-top: 10px;
+		height: 32px;
+		padding-left: 10px;
 		z-index: 1;
 
 		button {
-			margin-left: 10px;
-			padding: 5px 10px;
+			margin: 5px 10px 0 0;
+			padding: 2px 10px;
 			font-size: 12px;
 			border-radius: 20px;
 			font-weight: bold;
-			background: #111;
+			background: transparent;
 			color: #ddd;
-			border: 2px solid #333;
+			border: 2px solid #4a4a4a;
 			cursor: pointer;
 		}
 		button.active {
-			color: #50b3b7;
 			border-color: #50b3b7;
+			color: #fff;
 		}
+	}
+
+	.benchmark-box {
+		margin-top: 8px;
+		margin-left: 10px;
 	}
 
 	.exchange-dashboard {
@@ -95,6 +141,15 @@ const container = css`
 		overflow: hidden;
 		margin-right: -10px;
 	}
+
+	.pushed-margin {
+		margin-left: -10px;
+
+		th:first-child, td:first-child {
+			padding-left: 10px;
+		}
+	}
+	
 	.no-scroll-bar > div {
 		height: 100%;
 		position: absolute;
@@ -125,70 +180,158 @@ const container = css`
 		scrollbar-color: rgba(255,255,255,0.1) transparent;
 	}
 
-	
+	.websocket-status {
+		background: #dc3545;
+		position: fixed;
+    	width: 100%;
+	}
 `;
 
 class Exchange extends Component {
 	constructor(props) {
 		super(props)
 		this.state = {
-			chart: "tv"
+			chart: "tv",
+			toggle_trade: false,
+			dialog: undefined,
+			showBenchmark: true,
+			announcements: false
 		}
-		
+
+		this.eventUpdate = this.eventUpdate.bind(this)
+		this.Switchchart = this.Switchchart.bind(this)
+	}
+
+	componentDidMount() {
+		fetch(CONFIG.getEnv().ANNOUNCEMENT_JSON,{mode: "cors"}).then(e => e.json())
+        .then(data => {
+            const entries = data.entries
+            if (entries && entries.length > 0) {
+                this.setState({announcements: entries.slice(0,3)})
+            }
+		})
+
+		document.addEventListener('mouseenter', this.eventUpdate, false)
+	}
+
+	componentWillUnmount() {
+		document.removeEventListener('mouseenter', this.eventUpdate, false)
+	}
+	
+	eventUpdate() {
+		const { dispatch } = this.props
+		dispatch(reconnectIfNeeded())
+	}
+
+	resizeDepthChart() {
+		setTimeout(() =>{
+			window.depthChartWidget.setSize(null, null, false)
+		}, 0)
 	}
 
 	toggleChart(chart) {
 		this.setState({ chart: chart })
+		if(chart == "depth") {
+			this.resizeDepthChart()
+		}
+	}
+
+	toggleHistory() {
+		this.setState({toggle_trade: !this.state.toggle_trade})
+		this.resizeDepthChart()
+	}
+
+	toggleBenchmarkPrice() {
+		this.setState({ showBenchmark: !this.state.showBenchmark })
+	}
+
+	Switchchart() {
+		return(
+			<div className="switch-chart d-flex">
+				<button className={this.state.chart === "tv" ? "active": ""} onClick={() => this.toggleChart("tv")}>Price Chart</button>
+				<button className={this.state.chart === "depth" ? "active": ""} onClick={() => this.toggleChart("depth")}>Depth Chart</button>
+				{this.state.chart === "tv" ?
+				<React.Fragment>
+					<span className="benchmark-box">
+						<Switch label="Benchmark Price " onToggle={this.toggleBenchmarkPrice.bind(this)}  active={this.state.showBenchmark} dataTip="Enabling benchmark price aggregate prices from Binance/others<br/> to give more pricing information for you to make informed trades." />
+					</span>
+				</React.Fragment>
+				: null
+				}
+				
+			</div>
+		)
 	}
 
 	render() {
-		const Switchchart = () => {
-			return(
-				<div className="switch-chart d-flex">
-					<button className={this.state.chart === "tv" ? "active": ""} onClick={() => this.toggleChart("tv")}>TradingView</button>
-					<button className={this.state.chart === "depth" ? "active": ""} onClick={() => this.toggleChart("depth")}>Depth</button>
-				</div>
-			)
-		}
+		const { announcements } = this.state
+		const { websocket_status, publicKey } = this.props
 		return (
-		<div className={container + " container-fluid"}>
-			<div className="row flex-nowrap" style={{overflow:"hidden",minHeight:"calc(100vh - 120px)"}}>
-				<div className="exchange-left" style={{ display: this.props.leftOpen ? 'block': 'none'}}>
-					<OrderBook />
-				</div>
-				<div className="exchange-middle">
+			<div className={container + (announcements ? " has-announcement" : "") + (publicKey ? " has-user" : "")}>
+				<div className="d-flex">
 					<Header />
-						<Switchchart />
-						<Chart chartTools={true} className={this.state.chart === "tv" ? "d-block": "d-none"} />
-						<DepthChart  className={this.state.chart === "depth" ? "d-block": "d-none"} />
-					<div className="d-flex">
-						<Dashboard />
+					{publicKey ? <Menu /> : <Connect type="link" />}
+				
+					{ websocket_status ?
+						<div className="websocket-status text-center py-2 px-5">Connection Issue: {websocket_status}</div>
+						:null
+					}
+				</div>
+				{announcements ? <Announcement announcements={announcements}/> : null}
+				<div className="content d-flex">
+					<section className="compartment left-cols">
 						<Trade />
-					</div>
-					
-				</div>
-				<div className="exchange-right" style={{ display: this.props.rightOpen ? 'block' : 'none'}}>
-					<Menu />
-					<Leaderboard />
+						<hr/>
+						{publicKey ? <Balance /> : <Connect/>}
+					</section>
+					<section className="compartment left-cols">
+						<OrderBook />
+					</section>
 
-					<TradingHistory />
+					<div className="d-flex flex-column justify-content-between" style={{width: "calc(100% - 530px)"}}>
+						<div className="d-flex mb-2">
+							<div className="trade-toggle align-items-center">
+									<div className={"toggle cursor-pointer" + (this.state.toggle_trade ? " hide" : " show")}
+									onClick={() => this.toggleHistory()}>
+										{this.state.toggle_trade ? "HIDE" : "SHOW TRADE HISTORY"}
+									</div>
+								</div>
+							<section className="compartment" style={this.state.toggle_trade ? {width: "calc(100% - 270px)"} : {width: "100%"}}>
+								<this.Switchchart />
+								<Chart chartTools={true} showBenchmark={this.state.showBenchmark} className={this.state.chart === "tv" ? "d-block": "d-none" } />
+								<DepthChart className={this.state.chart === "depth" ? "d-block": "d-none"} />
+							</section>
+
+							<section className={"compartment cols" + (this.state.toggle_trade ? "" : " d-none")}>
+								<TradingHistory />
+							</section>
+						</div>
+						
+						{publicKey ? 
+							<section className="compartment">
+								<Orders />
+							</section>
+							: null
+						}
+					</div>
 				</div>
+
+				<section className="status">
+					<Status />
+				</section>
+				<ToastContainer />
 			</div>
-			<div className="row exchange-bottom">
-				<Orders />
-				<Status />
-			</div>
-			{ localStorage.getItem("firstTimeComplete") ? null : <FirstTime /> }
-		</div>
 		);
 	}
 }
 
 const mapStateToProps = (state) => ({
 		private_key: state.app.private_key,
+		publicKey: state.app.publicKey,
 		leftOpen: state.app.ui.leftOpen,
 		rightOpen: state.app.ui.rightOpen,
 		currentTicker: state.app.currentTicker,
+		websocket_status: state.app.websocket_status,
 	});
 
 
