@@ -1,12 +1,8 @@
 import React, { Component } from 'react';
 import Header from './headersimple.jsx';
-import QTTableView from './ui/tableView.jsx'
 import { connect } from 'react-redux'
 import { css } from 'emotion'
 import globalcss from './global-css.js'
-import { Link } from 'react-router-dom'
-import QTDeposit from './ui/deposit.jsx'
-import QTWithdraw from './ui/withdraw.jsx'
 import LeaderboardTable from './leaderboard.jsx'
 import MobileHeader from './ui/mobileHeader.jsx';
 
@@ -77,8 +73,8 @@ const container = css`
         .headline {
             display: inline-block;
             padding-left: 70px;
-            background: url('/public/images/trophy-blue.svg') no-repeat;
-            font-size: 22px;
+            background: url(${devicePath('public/images/trophy-blue.svg')}) no-repeat;
+            font-size: 18px;
         }
 
         .leaderboard-actions {
@@ -118,6 +114,9 @@ const container = css`
             padding: 0;
             border: 0;
         }
+        .leaderboard-share {
+            margin: 27px 25px 0 0;
+        }
         .content {
             padding: 15px;
         }
@@ -132,26 +131,12 @@ const container = css`
     }
 `
 
-const screenWidth = screen.width
-
 class Leaderboard extends Component {
-    constructor(props) {
-		super(props);
-		this.state = {
-			isMobile: screenWidth <= 992
-		};
-      }
-      
-    componentDidMount() {
-		if (!this.props.private_key) {
-			this.props.history.push("/login")
-        }
-    }
 
     render() {
         return(
-            <div className={container + (this.state.isMobile ? " mobile" : "")}>
-                {this.state.isMobile ?
+            <div className={container + (this.props.isMobile ? " mobile" : "")}>
+                {this.props.isMobile ?
                     <MobileHeader />
                     :
                     <div className="row header-row">
@@ -164,20 +149,20 @@ class Leaderboard extends Component {
                 </div>
 
                 <div className="content">
-                    <div className="banner">
-                        <div className="headline qt-font-light">
-                            Participate on the Paper Trading Contest <br/> and <b>win up to $50,000 USD*</b>
-                        </div>
-
-                        <div className="leaderboard-share d-flex justify-content-between float-right">
-                            <span>Share</span>
-                            <a><img src="/public/images/share/twitter.svg" /></a>
-                            <a><img src="/public/images/share/fbook.svg" /></a>
+                    <div className="banner d-flex justify-content-between">
+                        <div className="headline qt-font-light w-100">
+                            Currently, there is no active contest.<br/>
+                            We will notify you when the next contest is scheduled! 
                         </div>
                         <div className="leaderboard-actions float-right">
                             <a href="https://t.me/quantaexchange" target="_blank">Join Chat</a>
                             <a href="https://quantadex.com/fantasy" target="_blank">Read Rules</a>
 					    </div>
+                        <div className="leaderboard-share d-flex justify-content-between">
+                            <span>Share</span>
+                            <a className="ml-2"><img src={devicePath("public/images/share/twitter.svg")} /></a>
+                            <a className="ml-2"><img src={devicePath("public/images/share/fbook.svg")} /></a>
+                        </div>
                     </div>
                     <LeaderboardTable tableOnly={true} complete={true}/>
                 </div>
@@ -188,6 +173,7 @@ class Leaderboard extends Component {
 }
 
 const mapStateToProps = (state) => ({
+    isMobile: state.app.isMobile,
     private_key: state.app.private_key,
 });
 
