@@ -12,8 +12,7 @@ import globalcss from './global-css.js'
 const container = css`
 
 	.orderbook-ask, .orderbook-bid {
-		height: 35vh;
-		min-height:260px;
+		height: calc(50vh - 96px);
 		thead th{
 			position: -webkit-sticky;
 			position: sticky;
@@ -30,10 +29,18 @@ const container = css`
 	}
 
 	&.mobile {
-		padding: 0 15px;
+		padding: 0 10px;
 		.orderbook-ask, .orderbook-bid {
 			height: calc(50vh - 132px);
 			min-height: auto;
+		}
+
+		.perc-bar.left {
+			left: 0;
+		}
+
+		.perc-bar.right {
+			right: 0;
 		}
 	}
 `;
@@ -52,12 +59,16 @@ class OrderBook extends Component {
 	}
 
 	setAmount(side, e) {
+		const { mobile_nav } = this.props
+		if (mobile_nav) {
+			mobile_nav()
+		}
 		const set_time = new Date();
 		this.props.dispatch({
 			type: 'SET_AMOUNT',
 			data: {
 				inputBuy: e.price,
-				inputBuyAmount: e.amount,
+				inputBuyAmount: e.volume,
 				inputSide: side,
 				setTime: set_time
 			}
@@ -65,22 +76,21 @@ class OrderBook extends Component {
 	}
 
 	render() {
-		// const asksDataSource = lodash.takeRight(this.props.asks.dataSource,20).map((ask) => {
-		// 	return {
-		// 		...ask,
-		// 		price: parseFloat(ask.price).toFixed(this.props.decimals.value),
-		// 		total: parseFloat(ask.total).toFixed(this.props.decimals.maxTotalDecimals)
-		// 	}
-		// })
-		
-		var asksIterator = this.props.asks.dataSource.beginIterator();
+		const { currentTicker, asks, bids, decimals, mobile, mostRecentTrade, spread, mirror } = this.props
+		if (currentTicker == null) {
+			return <div></div>;
+		}
+		var asksIterator = asks.dataSource.beginIterator();
 		var asksDataSource = []
+		var askVolume = 0
 		while (asksIterator.value() !== null && asksDataSource.length < 20) {
 			const ask = asksIterator.value()
+			askVolume += parseFloat(ask.amount)
 			asksDataSource.push({
 				...ask,
-				price: parseFloat(ask.price).toFixed(this.props.decimals.value),
-				total: parseFloat(ask.total).toFixed(this.props.decimals.maxTotalDecimals)
+				price: parseFloat(ask.price).toFixed(decimals.value),
+				total: parseFloat(ask.total).toFixed(decimals.maxTotalDecimals),
+				volume: askVolume
 			})
 			asksIterator = asksIterator.next()
 		}
@@ -88,20 +98,89 @@ class OrderBook extends Component {
 		asksDataSource.reverse()
 
 
-		var bidsIterator = this.props.bids.dataSource.beginIterator();
+		var bidsIterator = bids.dataSource.beginIterator();
 		var bidsDataSource = []
+		var bidVolume = 0
 		while (bidsIterator.value() !== null && bidsDataSource.length < 20) {
 			const bid = bidsIterator.value()
+			bidVolume += parseFloat(bid.amount)
 			bidsDataSource.push({
 				...bid,
-				price: parseFloat(bid.price).toFixed(this.props.decimals.value),
-				total: parseFloat(bid.total).toFixed(this.props.decimals.maxTotalDecimals)
+				price: parseFloat(bid.price).toFixed(decimals.value),
+				total: parseFloat(bid.total).toFixed(decimals.maxTotalDecimals),
+				volume: bidVolume
 			})
 			bidsIterator = bidsIterator.next()
 		}
 
+		if (mirror) {
+			const bids_columns = [{
+				name: (ticker) => {return "Bids " + ticker.split('/')[0].split('0X')[0]},
+				key:"amount",
+				type:"number",
+				sortable:false,
+				color: (value) => {return "theme"},
+				fontSize:"base",
+				fontWeight:"light",
+				float:"left"
+			  },{
+				name: (ticker) => {return "Price " + ticker.split('/')[1].split('0X')[0]},
+				key:"price",
+				type:"number",
+				sortable:false,
+				color: (value) => {return "white"},
+				fontSize:"base",
+				fontWeight:"light",
+				float:"right pr-2"
+			  }]
+
+			const asks_columns = [{
+				name: (ticker) => {return "Price " + ticker.split('/')[1].split('0X')[0]},
+				key:"price",
+				type:"number",
+				sortable:false,
+				color: (value) => {return "white"},
+				fontSize:"base",
+				fontWeight:"light",
+				float:"left pl-2"
+			  },{
+				name: (ticker) => {return "Asks " + ticker.split('/')[0].split('0X')[0]},
+				key:"amount",
+				type:"number",
+				sortable:false,
+				color: (value) => {return "red"},
+				fontSize:"base",
+				fontWeight:"light",
+				float:"right"
+			  }]
+			return (
+				<div className={container + (mobile ? " mobile" : "")}>
+				<div className="qt-font-bold qt-font-normal my-3">ORDER BOOK</div>
+				<div className="d-flex">
+					<QTTableViewSimple key="bid_tv"
+						max={bids.max}
+						barColor="33,224,219"
+						barDir="right"
+						dataSource={bidsDataSource}
+						columns={bids_columns}
+						ticker={currentTicker}
+						onAction={this.setAmount.bind(this, 0)}
+					/>
+					<QTTableViewSimple key="ask_tv" dataSource={asksDataSource.reverse()} 
+						max={asks.max}
+						barColor="255,35,116"
+						barDir="left"
+						columns={asks_columns} 
+						ticker={currentTicker}
+						onAction={this.setAmount.bind(this, 1)}
+					/>
+				</div>
+				</div>
+			)
+		}
+
 		return (
-			<div className={container + (this.props.mobile ? " mobile" : "")}>
+			<div className={container + (mobile ? " mobile" : "")}>
 				<section className="orderbook-title">
 					<div className="d-flex justify-content-between align-items-center">
 						<div className="qt-font-bold qt-font-normal">ORDER BOOK</div>
@@ -114,30 +193,35 @@ class OrderBook extends Component {
 							onChange={this.updateDigits}/> */}
 					</div>
 				</section>
-				<section className="orderbook-ask no-scroll-bar">
+				<section className="orderbook-ask no-scroll-bar pushed-margin">
 					<div id="ask-section">
 						<QTTableViewSimple key="ask_tv" dataSource={asksDataSource} 
-							columns={this.props.asks.columns} 
-							ticker={this.props.currentTicker}
+							max={asks.max}
+							barColor="255,35,116"
+							barDir="left"
+							columns={asks.columns} 
+							ticker={currentTicker}
 							onAction={this.setAmount.bind(this, 1)}
 						/>
 					</div>
 				</section>
 				<section className="orderbook-middle d-flex justify-content-between">
 					<div className="d-flex flex-column justify-content-center">
-						<div className="qt-color-theme qt-font-huge qt-font-light">{this.props.mostRecentTrade.price}</div>
-						{/* <div className="qt-number-normal qt-opacity-64">{this.props.spreadDollar}</div> */}
+						<div className="qt-color-theme qt-font-huge qt-font-light">{mostRecentTrade.price}</div>
 					</div>
 					<div className="d-flex flex-column justify-content-center">
 						<div className="qt-opacity-half qt-font-base text-right">Spread</div>
-						<div className="qt-number-small text-right">{this.props.spread != undefined ? this.props.spread.toFixed(2) + "%" : "N/A"}</div>
+						<div className="qt-number-small text-right">{spread != undefined ? spread.toFixed(2) + "%" : "-"}</div>
 					</div>
 				</section>
-				<section className="orderbook-bid no-scroll-bar">
+				<section className="orderbook-bid no-scroll-bar pushed-margin">
 					<div>
 						<QTTableViewSimple key="bid_tv"
+							max={bids.max}
+							barColor="33,224,219"
+							barDir="left"
 							dataSource={bidsDataSource}
-							columns={this.props.bids.columns}
+							columns={bids.columns}
 							HideHeader={true}
 							onAction={this.setAmount.bind(this, 0)}
 						/>
@@ -149,6 +233,7 @@ class OrderBook extends Component {
 }
 
 const mapStateToProps = (state) => ({
+	orderbook: state.app.orderBook, 
   	bids: state.app.orderBook.bids,
   	asks: state.app.orderBook.asks,
 	decimals: state.app.orderBook.decimals,
