@@ -1,30 +1,37 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux'
 import moment from 'moment';
-import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
+import { toast } from 'react-toastify';
 import { css } from 'emotion'
 import globalcss from './global-css.js'
 
 import QTTabBar from './ui/tabBar.jsx'
 import QTButton from './ui/button.jsx'
 import QTTableViewSimple from './ui/tableViewSimple.jsx'
+import Loader from './ui/loader.jsx'
+import Switch from "./ui/switch.jsx"
 
-import {cancelTransaction} from "../redux/actions/app.jsx";
+import { cancelTransaction, loadOrderHistory, TOGGLE_CONNECT_DIALOG, updateUserData } from "../redux/actions/app.jsx";
 import ReactGA from 'react-ga';
+import lodash from 'lodash';
 
 
 const container = css`
   position: relative;
+  height: 260px;
   width: 100%;
-  padding: 20px 35px 0;
-  margin-bottom: 54px;
+  padding: 10px;
 
   .order-list {
-    margin: 15px 0;
+    margin-top: 4px;
+    height: 205px;
+    overflow: hidden;
+    overflow-y: scroll;
+
     table {
-      tbody tr {
-        border-top: 1px solid #333;
+      tr {
+        border-bottom: 1px solid #333;
+        cursor: default;
       }
       tbody td {
         padding: 4px 0;
@@ -92,7 +99,7 @@ const container = css`
     z-index: 2;
   }
   thead th {
-    top: 28px;
+    top: 0;
     z-index: 1;
   }
 
@@ -100,7 +107,7 @@ const container = css`
     position: absolute;
     right: 15px;
     top: 4px;
-    background: url(/public/images/up-arrow.svg) repeat-y 100%;
+    background: url(${devicePath("public/images/up-arrow.svg")}) repeat-y 100%;
     height: 18px;
     padding-right: 20px;
     line-height: 20px;
@@ -114,39 +121,18 @@ const container = css`
     font-size: 16px;
     color: #777;
   }
-  
-  .loader {
-    display: none;
-    border: 2px solid rgba(255,255,255,0.5);
-    border-radius: 50%;
-    border-top: 2px solid #fff;
-    width: 15px;
-    height: 15px;
-    float: right;
-    margin-right: 28px;
-    -webkit-animation: spin 2s linear infinite;
-    animation: spin 2s linear infinite;
-  }
-  
-  @-webkit-keyframes spin {
-      0% { -webkit-transform: rotate(0deg); }
-      100% { -webkit-transform: rotate(360deg); }
-  }
-      
-  @keyframes spin {
-      0% { transform: rotate(0deg); }
-      100% { transform: rotate(360deg); }
-  }
 
   &.mobile {
     padding: 15px;
     margin: 0;
-    .scroll-up {
-      display: none;
+    
+    .sticky, thead th {
+      background: #0A121E;
     }
 
     .order-list {
       padding: 0;
+      height: calc(100vh - 168px);
 
       .list-row {
         padding: 5px 0;
@@ -154,16 +140,27 @@ const container = css`
       }
 
       .item-assets {
-        width: 30%;
+        width: 50%;
       }
       .item-price {
         width:50%;
       }
-
-      .item-type-BUY, .item-type-SELL {
+      .item-type {
         width: 20%;
         padding-right: 20px;
-        background: url("../public/images/menu-arrow-down.svg") no-repeat 100% 50%;
+      }
+      .item-status {
+        width: 30%;
+      }
+
+      .filled-order {
+        .item-price {
+          width: 30%;
+        }
+      }
+
+      .item-type-BUY, .item-type-SELL {
+        background: url(${devicePath("public/images/menu-arrow-down.svg")}) no-repeat 100% 50%;
       }
 
       .item-type-BUY {
@@ -191,9 +188,6 @@ const container = css`
           padding: 0 10px;
         }
       }
-      .item-details.active {
-        display: block;
-      }
     }
   }
 `;
@@ -203,64 +197,71 @@ class Orders extends Component {
     super(props);
     this.state = {
       selectedTabIndex: 0,
+      hideMarkets: false,
       isFocused: false,
-      cancelling: [],
-      selectedRow: null
+      cancelling: {},
+      selectedRow: null,
+      loading: false,
+      page: 1,
     };
+
+    this.OpenOrders = this.OpenOrders.bind(this)
+    this.FilledOrders = this.FilledOrders.bind(this)
   }
 
-  goToTop() {
-    window.scrollTo(0,0);
-    this.setState({isFocused: false})
-  }
-
-  handleSwitch(index, selected) {
+  handleSwitch(index) {
     this.setState({selectedTabIndex: index})
-    if(index == selected && this.state.isFocused) {
-      this.goToTop()
-    } else {
-      window.scrollTo(0,document.body.scrollHeight);
-      this.setState({isFocused: true})
-    }
+    this.props.dispatch(updateUserData())
+    document.getElementById("scroll-order-list").scrollTop = 0
   }
 
   notify_success = (toastId) => toast.update(toastId, {
     render: "Order cancelled",
     type: toast.TYPE.SUCCESS,
-    autoClose: 5000,
-    position: toast.POSITION.TOP_CENTER
+    autoClose: 2000,
+    position: toast.POSITION.TOP_CENTER,
+    pauseOnFocusLoss: false,
+    pauseOnHover: false
   });
   notify_failed = (toastId) => toast.update(toastId, {
     render: "Unable to cancel order",
     type: toast.TYPE.ERROR,
-    autoClose: 5000,
-    position: toast.POSITION.TOP_CENTER
+    autoClose: 2000,
+    position: toast.POSITION.TOP_CENTER,
+    pauseOnFocusLoss: false,
+    pauseOnHover: false
   });
 
   handleCancel(market, order) {
     const toastId = toast("CANCELING...", { autoClose: false, position: toast.POSITION.TOP_CENTER });
-    if (!this.props.mobile) {
-      var id = order.replace(/\./g, '-')
-      document.querySelectorAll('#cancel-' + id + ' button')[0].style.display = 'none';
-      document.querySelectorAll('#cancel-' + id + ' .loader')[0].style.display = 'block';
-    }
+    const self = this
+    this.setState((state) => {
+      let cancelling = state.cancelling
+      cancelling[order] = true
+      return {cancelling}
+    })
 
     ReactGA.event({
       category: 'CANCEL',
       action: market
     });
-    this.props.dispatch(cancelTransaction(market, order))
+    this.props.dispatch(cancelTransaction(order))
     .then((e) => {
-      if (!this.props.mobile) {
-        document.querySelectorAll('#cancel-' + id + ' .loader')[0].style.display = 'none';
-      }
+      setTimeout(() => {
+        self.setState((state) => {
+          let cancelling = state.cancelling
+          delete cancelling[order]
+          return {cancelling}
+        })
+      }, 5000)
       
       this.notify_success(toastId)
     }).catch((e) => {
-      if (!this.props.mobile) {
-        document.querySelectorAll('#cancel-' + id + ' button')[0].style.display = 'inherit';
-        document.querySelectorAll('#cancel-' + id + ' .loader')[0].style.display = 'none';
-      }
+      this.setState((state) => {
+        let cancelling = state.cancelling
+        delete cancelling[order]
+        return {cancelling}
+      })
       this.notify_failed(toastId)
     })
   }
@@ -273,104 +274,178 @@ class Orders extends Component {
     }
   }
 
-  render() {
-    // const orders_data = [
-    //   ["BNCBTC","0.234567 BTC","0.234567 BTC","0.234567 BTC","0.234567 BTC","BUY","Filled","12 JAN, 12:34:15"],
-    //   ["BNCBTC","0.234567 BTC","0.234567 BTC","0.234567 BTC","0.234567 BTC","BUY","Filled","12 JAN, 12:34:15"]
-    // ]
-    const tabs = {
-      names: ['ACTIVE ORDERS', 'FILLED ORDERS'],
-      selectedTabIndex: 0,
+  handleScroll = lodash.throttle((e) => {
+    // console.log(this.props)
+    if (this.state.loading) {
+      return
     }
-    
-    const OrdersList = () => {
-      if (this.state.selectedTabIndex == 0) {
-        if (this.props.openOrders.dataSource.length == 0) {
-          return <div className="empty-list">You have no active orders</div>
-        }
-        if (this.props.mobile) {
-          return (
-            <div>
-              {this.props.openOrders.dataSource.map(row => {
-                return (
-                  <div key={row.id} className="list-row" onClick={() => this.toggleDetails(row.id)}>
-                    <div className="d-flex list-item">
-                      <span className="item-assets">{row.assets}</span>
-                      <span className="item-price text-right">{row.price}</span>
-                      <span className={"text-right item-type-" + row.type}>{row.type}</span>
-                    </div>
-                    <div className={"item-details" + (this.state.selectedRow == row.id ? " active" : "")}>
-                      <span className="item"><span className="label">AMOUNT</span> {row.amount}</span>
-                      <span className="item"><span className="label">TOTAL</span> {row.total}</span>
-                      <button onClick={() => this.handleCancel(row.assets, row.id)}>CANCEL</button>
-                    </div>
-                  </div>
-                )
-              })}
-            </div>
-          )
-        } else {
-          return (
-            <QTTableViewSimple dataSource={this.props.openOrders.dataSource} columns={this.props.openOrders.columns}
-            cancelOrder={this.handleCancel.bind(this)} />
-          )
-        }
-          
-      } else {
-        if (this.props.filledOrders.dataSource.length == 0) {
-          return <div className="empty-list">You have no recent filled orders in this market</div>
-        }
 
-        if (this.props.mobile) {
-          return (
-            <div>
-              {this.props.filledOrders.dataSource.map(row => {
-                return (
-                  <div key={row.id} className="list-row" onClick={() => this.toggleDetails(row.id)}>
-                    <div className="d-flex list-item">
-                      <span className="item-assets">{row.assets}</span>
-                      <span className="item-price text-right">{row.price}</span>
-                      <span className={"text-right item-type-" + row.type}>{row.type}</span>
-                    </div>
-                    <div className={"item-details" + (this.state.selectedRow == row.id ? " active" : "")}>
-                      <span className="item"><span className="label">AMOUNT</span> {row.amount}</span>
-                      <span className="item"><span className="label">TOTAL</span> {row.total}</span>
-                    </div>
-                  </div>
-                )
-              })}
-            </div>
-          )
-        } else {
-          return (
-            <QTTableViewSimple dataSource={this.props.filledOrders.dataSource} columns={this.props.filledOrders.columns}/>
-          )
+    if (this.state.selectedTabIndex !== 0) {
+      if (e.scrollTop > e.scrollHeight - 300) {
+        if (this.state.selectedTabIndex == 1) {
+          if (this.props.filledOrders.end) {
+            return
+          }
+          
+          const page = this.state.page
+          this.setState({loading: true})
+          this.props.dispatch(loadOrderHistory(page)).then(() => {
+            this.setState({loading: false, page: page + 1})
+          })
         }
       }
     }
+  }, 200)
+
+  openConnect() {
+    const { mobile_nav, dispatch } = this.props
+        if (mobile_nav) {
+            mobile_nav("connect")
+        } else {
+            dispatch({
+                type: TOGGLE_CONNECT_DIALOG,
+                data: "connect"
+            })
+        }
+  }
+
+  OpenOrders() {
+    const { openOrders, currentTicker, mobile, private_key } = this.props
+    const { cancelling } = this.state
+
+    if (openOrders.dataSource.length == 0) {
+      return <div className="empty-list">You have no active orders</div>
+    }
+
+    const dataSource = openOrders.dataSource.filter(row => !this.state.hideMarkets || row.assets === currentTicker)
+    dataSource.sort((a,b) => (a.id > b.id) ? -1 : ((b.id > a.id) ? 1 : 0))
+
+    if (mobile) {
+      return (
+        <div>
+          <div className="list-row sticky">
+            <div className="d-flex list-item qt-white-27">
+              <span className="item-assets">PAIR</span>
+              <span className="item-price text-right">PRICE</span>
+              <span className="item-type text-right">TYPE</span>
+            </div>
+          </div>
+          {dataSource.map((row) => {
+            return (
+              <div key={row.id} className="list-row" onClick={() => this.toggleDetails(row.id)}>
+                <div className="d-flex list-item">
+                  <span className="item-assets">{row.pair}</span>
+                  <span className="item-price text-right">{row.price}</span>
+                  <span className={"text-right item-type item-type-" + row.type}>{row.type}</span>
+                </div>
+                <div className={"item-details justify-content-between flex-wrap mt-2" + (this.state.selectedRow == row.id ? " d-flex" : "")}>
+                  <span className="item"><span className="label">AMOUNT</span> {row.amount}</span>
+                  <span className="item"><span className="label">TOTAL</span> {row.total}</span>
+                  {
+                    cancelling[row.id] ?
+                    <Loader size={15} />
+                    :
+                    <button onClick={() => !private_key ? this.openConnect() : this.handleCancel(row.assets, row.id)}>
+                      CANCEL
+                    </button>
+                  }
+                </div>
+              </div>
+            )
+          })}
+        </div>
+      )
+    } else {
+      return (
+        <QTTableViewSimple dataSource={dataSource} 
+        columns={openOrders.columns}
+        disabled={!private_key} 
+        openConnect={this.openConnect.bind(this)}
+        cancelling={cancelling}
+        cancelOrder={this.handleCancel.bind(this)} />
+      )
+    }
+  }
+
+  FilledOrders() {
+    if (this.props.filledOrders.dataSource.length == 0) {
+      return <div className="empty-list">You have no recent filled orders in this market</div>
+    }
+
+    const dataSource = this.props.filledOrders.dataSource.concat(this.props.filledOrders.dataSource2).filter(row => !this.state.hideMarkets || row.assets === this.props.currentTicker)
+
+    if (this.props.mobile) {
+      return (
+        <div className="filled-orders">
+          <div className="list-row sticky">
+            <div className="d-flex list-item qt-white-27">
+              <span className="item-assets">PAIR</span>
+              <span className="item-price text-right">PRICE</span>
+              <span className="item-status text-right">STATUS</span>
+              <span className="item-type text-right">TYPE</span>
+            </div>
+          </div>
+          {dataSource.map((row, index) => {
+            return (
+              <div key={row.id + index} className="list-row" onClick={() => this.toggleDetails(index)}>
+                <div className="d-flex list-item">
+                  <span className="item-assets">{row.pair}</span>
+                  <span className="item-price text-right">{row.price}</span>
+                  <span className="item-status text-right">{row.status}</span>
+                  <span className={"text-right item-type item-type-" + row.type}>{row.type}</span>
+                </div>
+                <div className={"item-details justify-content-between flex-wrap mt-2" + (this.state.selectedRow == index ? " d-flex" : "")}>
+                  <span className="item"><span className="label">AMOUNT</span> {row.amount}</span>
+                  <span className="item"><span className="label">TOTAL</span> {row.total}</span>
+                </div>
+              </div>
+            )
+          })}
+        </div>
+      )
+    } else {
+      return (
+        <QTTableViewSimple dataSource={dataSource} columns={this.props.filledOrders.columns}/>
+      )
+    }
+  }
+
+  render() {
+    const { mobile } = this.props
+    const tabs = {
+      names: ['Open Orders', 'Order History'],
+      selectedTabIndex: 0,
+    }
+    const OrdersList = [<this.OpenOrders />, <this.FilledOrders />]
 
     return (
-      <div className={container + (this.props.mobile ? " mobile" : "")}>
-      <div className="sticky">
-          <div className="scroll-up" onClick={this.goToTop.bind(this)}>SCROLL UP</div>
+      <div className={container + (mobile ? " mobile" : "")}>
+      <div className="sticky d-flex align-items-center">
           <QTTabBar
-            className="underline small static set-width qt-font-bold d-flex justify-content-center"
-            width={200}
+            className={"small static set-width qt-font-bold d-flex justify-content-left " + (mobile ? "underline" : "button")}
+            width={120}
             gutter={10}
             tabs = {tabs}
             switchTab = {this.handleSwitch.bind(this)}
           />
+          <Switch label="Current Market Only" 
+            className={mobile ? "ml-4" : ""}
+            active={this.state.hideMarkets} 
+            onToggle={() => this.setState({hideMarkets: !this.state.hideMarkets})} />
         </div>
-        <section className="order-list container-fluid">
-          <OrdersList />
+        <section className="order-list no-scroll-bar">
+          <div id="scroll-order-list" onScroll={(e) => this.handleScroll(e.target)}>
+            {OrdersList[this.state.selectedTabIndex]}
+            {this.state.loading && <Loader margin="10px auto"/>}
+          </div>
 				</section>
-        <ToastContainer />
       </div>
     )
   }
 }
 
 const mapStateToProps = (state) => ({
+    private_key: state.app.private_key,
   	bids: state.app.tradeBook.bids,
   	asks: state.app.tradeBook.asks,
 		currentPrice: state.app.currentPrice,
