@@ -3,25 +3,26 @@ import { connect } from 'react-redux'
 import moment from 'moment';
 
 import { css } from 'emotion'
-import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
+import { toast } from 'react-toastify';
 import globalcss from './global-css.js'
 
 import QTTabBar from './ui/tabBar.jsx'
 import QTDropdown from './ui/dropdown.jsx'
 import QTButton from './ui/button.jsx'
-import {Token, SmallToken} from './ui/ticker.jsx'
+import { Token, SmallToken } from './ui/ticker.jsx'
 import Loader from './ui/loader.jsx'
+import {LockIcon} from './ui/account_lock.jsx'
+import Utils from '../common/utils'
+import lodash from 'lodash';
 
-import {buyTransaction} from "../redux/actions/app.jsx";
-import {sellTransaction} from "../redux/actions/app.jsx";
+import { buyTransaction, sellTransaction, TOGGLE_CONNECT_DIALOG } from "../redux/actions/app.jsx";
 import ReactGA from 'react-ga';
 
 const container = css`
   font-family: SFCompactTextRegular;
-  width: 360px;
+  width: 100%;
   display: inline-block;
-  padding: 20px;
+  padding: 0;
   
   .buy-sell-toggle {
     width: 100%;
@@ -57,9 +58,16 @@ const container = css`
       border-radius: 2px;
       cursor: pointer;
     }
-    .inactive {
-      display: none;
+    button:disabled {
+      background-color: #31383d;
     }
+  }
+
+  button.connect-btn {
+    background-color: transparent;
+    border: 2px solid #66d7d7;
+    border-radius: 5px;
+    color: #66d7d7;
   }
 
   .buy-btn, .sell-btn {
@@ -74,19 +82,38 @@ const container = css`
 
   .input-container {
     display: flex;
-    margin: 15px 0;
+    align-items: center;
+    margin: 6px 0;
+
     label {
-      flex: 0 0 70px;
-      line-height: 32px;
+      flex: 1 1 60px;
+      margin: 0;
     }
+
     input {
       border-radius: 2px;
-      padding-right: 40px;
+      padding: 0 40px 0 10px;
+      text-align: left;
     }
     input::selection {
       background-color: #fff;
       color: #333;
     }
+    input:invalid {
+      box-shadow: none;
+    }
+    
+    .percent-container, input {
+      flex: 1 1 calc(100% - 60px);
+    }
+
+    .amount-select {
+      line-height: 17px; 
+      width: 23%;
+      border: 1px solid rgba(255,255,255,0.27);
+      border-radius: 2px;
+    }
+
     span {
       position: absolute;
       right: 10px;
@@ -186,238 +213,391 @@ const container = css`
     border: solid 2px #1cdad8;
   }
 
+  .fees-container {
+      table {
+          width: 100%;
+          tr td:first-child, tr td:last-child {
+              width: 1%;
+          }
+      }
+  }
+
+  .rounded-0-left {
+    border-top-left-radius: 0 !important;
+    border-bottom-left-radius: 0 !important;
+  }
+
   &.mobile {
     width: 100%;
+    padding: 10px;
   }
 `;
 
 class Trade extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-        processing: false,
-        qty: 0.05,
-        price: 1,
-        inputSetTime: 0
-      };
-  }
-
-	componentWillReceiveProps(nextProps) {
-    if (nextProps.inputSetTime != undefined && nextProps.inputSetTime != this.state.inputSetTime) {
-      this.setState({
-        qty: nextProps.inputBuyAmount,
-        price: nextProps.inputBuy,
-        inputSetTime: nextProps.inputSetTime
-      })
-      this.switchTradeTo(nextProps.inputSide)
-    }
-	}
-
-  notify_success = (toastId, msg) => toast.update(toastId, {
-    render: msg,
-    type: toast.TYPE.SUCCESS,
-    autoClose: 5000,
-    position: toast.POSITION.TOP_CENTER
-  });
-  notify_failed = (toastId, msg) => toast.update(toastId, {
-    render: msg,
-    type: toast.TYPE.ERROR,
-    autoClose: 5000,
-    position: toast.POSITION.TOP_CENTER
-  });
-
-  toastMsg(label, success, e) {
-    const msg = ( <div>
-      <span>{label}</span><br/>
-      <span>{success ? "OrderId: " + e.id.substr(0,10) : 
-                      "Failed order: " +  (e.message.includes("insufficient balance") ? "Insufficient Balance" : "Unable to place order")}</span>
-      </div> )
-    return msg
-  }
-
-	handleBuy(e) {
-    const label = this.props.currentTicker.split('/')[1] + " " + this.state.price + " @ " + this.state.qty
-    const toastId = toast("BUYING " + label, { autoClose: false, position: toast.POSITION.TOP_CENTER });
-
-    ReactGA.event({
-      category: 'BUY',
-      action: this.props.currentTicker
-    });
-
-    this.setState({processing: true})
-    this.props.dispatch(buyTransaction(this.props.currentTicker, this.state.price, this.state.qty))
-    .then((e) => {
-      const msg = this.toastMsg("BUY " + label, true, e)
-      this.notify_success(toastId, msg)
-    }).catch((e) => {
-      const msg = this.toastMsg("BUY " + label, false, e)
-      this.notify_failed(toastId, msg)
-    }).finally(() => {
-      this.setState({processing: false})
-    })
-	}
-
-	handleSell(e) {
-    const label = this.props.currentTicker.split('/')[1] + " " + this.state.price + " @ " + this.state.qty
-    const toastId = toast("SELLING " + label, { autoClose: false, position: toast.POSITION.TOP_CENTER });
-
-    ReactGA.event({
-      category: 'SELL',
-      action: this.props.currentTicker
-    });
-
-    this.setState({processing: true})
-    this.props.dispatch(sellTransaction(this.props.currentTicker, this.state.price, this.state.qty))
-    .then((e) => {
-      const msg = this.toastMsg("SELL " + label, true, e)
-      this.notify_success(toastId, msg)
-    }).catch((e) => {
-      const msg = this.toastMsg("SELL " + label, false, e)
-      this.notify_failed(toastId, msg)
-    }).finally(() => {
-      this.setState({processing: false})
-    })
-	}
-
-	handlePriceInputChange(e) {
-         this.setState({
-           price: e.target.value
-         });
-   	}
-
-	handleQtyInputChange(e) {
-         this.setState({
-           qty: e.target.value
-         });
-  }
-
-  handleInputFocus(e) {
-    e.target.select();
-  }
-
-  switchTradeTo(side) {
-    var to_hide = side === 1 ? document.getElementsByClassName('buy-btn') : 
-                                                  document.getElementsByClassName('sell-btn');
-    var to_show = side === 1 ? document.getElementsByClassName('sell-btn') : 
-                                                  document.getElementsByClassName('buy-btn');
-
-    for (let i=0; i < to_hide.length; i++) {
-      to_hide[i].classList.add("inactive");
-    }
-    for (let i=0; i < to_show.length; i++) {
-      to_show[i].classList.remove("inactive");
-    }
-  }
-
-  render() {
-    const tabs = {
-      names: ['LIMIT','MARKET','STOP-LIMIT'],
-      selectedTabIndex: 0,
+    constructor(props) {
+        super(props);
+        this.state = {
+            trade_side: this.props.trade_side || 0,
+            processing: false,
+            qty: 0,
+            price: 0,
+            total: 0,
+            price_set: false,
+            inputSetTime: 0,
+            currentPrice: undefined,
+            currentTicker: undefined
+        };
     }
 
-    const dropdown_items = {
-        items: ["10%","20%","25%","50%","75%","90%","100%"],
-        value: "20%"
-    }
-
-    const tradingPair = this.props.currentTicker.split('/')
-    const balance = this.props.balance
-    var pairBalance = []
-		Object.keys(balance).forEach((currency) => {
-			var item = {}
-      item.currency = window.assets[balance[currency].asset].symbol
-      if (!tradingPair.includes(item.currency)) {
+    componentDidMount() {
+      const { currentTicker, mobile, inputSetTime, currentPrice, inputBuyAmount, inputBuy} = this.props
+      if (!mobile) {
         return
       }
-			item.amount = balance[currency].balance ? balance[currency].balance : 0
-			pairBalance.push(item)
-    })
-    
-    return (
-      <div className={container + " container-fluid" + (this.props.mobile ? " mobile" : "")}>
-        <div className="buy-sell-toggle">
-          <button id="buy-switch" className="buy-btn" onClick={this.switchTradeTo.bind(this, 0)}>BUY {tradingPair[0]}</button>
-          <button id="sell-switch" className="sell-btn inactive" onClick={this.switchTradeTo.bind(this, 1)}>SELL {tradingPair[0]}</button>
-        </div>
+      if (inputSetTime && currentPrice.price && (new Date() - inputSetTime) < 50) {
+        this.setState({
+          qty: parseFloat(inputBuyAmount),
+          price: parseFloat(inputBuy),
+          total: parseFloat(inputBuyAmount) * parseFloat(inputBuy),
+          inputSetTime: inputSetTime,
+          currentPrice: currentPrice.price,
+          currentTicker: currentTicker
+        })
+      } else {
+        this.setState({
+          price: parseFloat(currentPrice.price),
+          currentPrice: currentPrice.price,
+          currentTicker: currentTicker
+        })
+      }
+    }
 
-        <div className="transac-actions">
-          
-            <div className="input-container">
-              <label>PRICE</label>
-              <input type="number" className="trade-input qt-number-bold qt-font-small"
-                    name="price"
-                    onFocus={this.handleInputFocus.bind(this)}
-                    min="0"
-                    step="0.0000001"
-                     value={this.state.price}
-                     onChange={this.handlePriceInputChange.bind(this)} />
-              <SmallToken name={tradingPair[1]}/>
-            </div>
-            <div className="input-container">
-               <label>AMOUNT {/*<Token name={tradingPair[0]}/>*/}</label> 
-              {/* <QTDropdown
-									items={dropdown_items.items}
-									value={dropdown_items.value}
-									className="up bordered dark qt-font-base qt-font-bold"
-									reverse={true}
-									width="50"
-									height="32"
-									onChange={() => {console.log("dropdown changed value")}}/> */}
-              <input type="number" className="trade-input qt-number-bold qt-font-small"
-                      name="amount"
-                      onFocus={this.handleInputFocus.bind(this)}
-                      min="0"
-                      step="0.0000001"
-                       value={this.state.qty}
-                       onChange={this.handleQtyInputChange.bind(this)} />
-              <SmallToken name={tradingPair[0]}/>
-            </div>
-            <div className="input-container">
-               <label>TOTAL {/*<Token name={tradingPair[1]}/> */}</label>
-              <input
-										type="number"
-                    className="trade-input qt-number-bold qt-font-small"
-                    name="total"
-                    min="0"
-                    step="0.0000001"
-                    onFocus={this.handleInputFocus.bind(this)}
-										value={((this.state.qty*10000000) * (this.state.price*10000000))/100000000000000}
-									 />
-              <SmallToken name={tradingPair[1]}/>
-            </div>
+    componentWillReceiveProps(nextProps) {
+        const { inputSetTime, currentPrice, currentTicker } = this.state
+        if (nextProps.inputSetTime != undefined && nextProps.inputSetTime != inputSetTime) {
+            this.setState({
+                qty: parseFloat(nextProps.inputBuyAmount),
+                price: parseFloat(nextProps.inputBuy),
+                total: parseFloat(nextProps.inputBuyAmount) * parseFloat(nextProps.inputBuy),
+                inputSetTime: nextProps.inputSetTime
+            })
+        }
+        if (currentPrice !== nextProps.currentPrice.price && (currentTicker !== nextProps.currentPrice.ticker || !currentPrice)) {
+          this.setState({
+            currentTicker: nextProps.currentPrice.ticker, 
+            currentPrice: nextProps.currentPrice.price, 
+            price: nextProps.currentPrice.price, 
+            qty: 0,
+            total: 0
+          })
+        }
+    }
 
-          <button id="sell-action" className="sell-btn inactive" onClick={this.handleSell.bind(this)}>{this.state.processing ? <Loader /> : "PLACE SELL ORDER"}</button>
-          <button id="buy-action" className="buy-btn" onClick={this.handleBuy.bind(this)}>{this.state.processing ? <Loader /> : "PLACE BUY ORDER"}</button>
-        </div>
-        <div className="d-flex justify-content-center align-items-center qt-font-small">
-						{
-							pairBalance.map((item) => {
-                
-								return (
-									<span className="trade-balance">{item.currency} Balance: {item.amount}</span>
-								)
-							})
-						}
-          </div>
-          <ToastContainer />
-      </div>
-    )
-  }
+    notify_success = (toastId, msg) => toast.update(toastId, {
+        render: msg,
+        type: toast.TYPE.SUCCESS,
+        autoClose: 2000,
+        position: toast.POSITION.TOP_CENTER,
+        pauseOnFocusLoss: false,
+        pauseOnHover: false
+    });
+    notify_failed = (toastId, msg) => toast.update(toastId, {
+        render: msg,
+        type: toast.TYPE.ERROR,
+        autoClose: 2000,
+        position: toast.POSITION.TOP_CENTER,
+        pauseOnFocusLoss: false,
+        pauseOnHover: false
+    });
+
+    toastMsg(label, success, e) {
+        const msg = (<div>
+            <span>{label}</span><br />
+            <span>{success ? "Order ID: " + e.trx.operation_results[0][1] :
+                "Failed order: " + (e.message.includes("insufficient balance") ? "Insufficient Balance" : "Unable to place order: " + e.message)}</span>
+        </div>)
+        return msg
+    }
+
+    handleBuy(e) {
+        const { dispatch, currentTicker } = this.props
+        const { price, qty } = this.state
+        const asset = currentTicker.split('/')[0].split('0X')
+        const label = asset[0] + (asset[1] ? '0x' + asset[1].substr(0,4) : "") + " " + qty + " @ " + price
+        const toastId = toast("BUYING " + label, { autoClose: false, position: toast.POSITION.TOP_CENTER });
+
+        ReactGA.event({
+            category: 'BUY',
+            action: currentTicker
+        });
+
+        this.setState({ processing: true })
+        dispatch(buyTransaction(currentTicker, price, qty))
+            .then((e) => {
+                const msg = this.toastMsg("BUY " + label, true, e)
+                this.notify_success(toastId, msg)
+            }).catch((e) => {
+                const msg = this.toastMsg("BUY " + label, false, e)
+                this.notify_failed(toastId, msg)                
+                // log here
+            }).finally(() => {
+                this.setState({ processing: false })
+            })
+    }
+
+    handleSell(e) {
+        const { dispatch, currentTicker } = this.props
+        const { price, qty } = this.state
+        const asset = currentTicker.split('/')[0].split('0X')
+        const label = asset[0] + (asset[1] ? '0x' + asset[1].substr(0,4) : "") + " " + qty + " @ " + price
+        const toastId = toast("SELLING " + label, { autoClose: false, position: toast.POSITION.TOP_CENTER });
+
+        ReactGA.event({
+            category: 'SELL',
+            action: currentTicker
+        });
+
+        this.setState({ processing: true })
+        dispatch(sellTransaction(currentTicker, price, qty))
+            .then((e) => {
+                const msg = this.toastMsg("SELL " + label, true, e)
+                this.notify_success(toastId, msg)
+            }).catch((e) => {
+                const msg = this.toastMsg("SELL " + label, false, e)
+                this.notify_failed(toastId, msg)
+            }).finally(() => {
+                this.setState({ processing: false })
+            })
+    }
+
+    switchTradeTo(side) {
+        this.setState({trade_side: side})
+    }
+
+    setPercentAmount(perc, symbol) {
+      const { balance } = this.props
+      const { trade_side, price} = this.state
+      const coin = window.assetsBySymbol[symbol].id
+      const amount = balance[coin] ? balance[coin].balance * (parseInt(perc)/100) : 0
+
+      let qty, total
+      if (trade_side == 0) {
+        qty = amount / price
+        total = amount
+      } else {
+        qty = amount
+        total = qty * price
+      }
+      this.setState({qty, total})
+    }
+
+    render() {
+        const { currentTicker, balance, mobile, mobile_nav, fee, publicKey, private_key } = this.props
+        const { trade_side, price, qty, total, processing} = this.state
+        const dropdown_items = {
+            items: ["25%", "50%", "75%", "100%"],
+            value: "25%"
+        }
+        
+        if (currentTicker == null) {
+          return <div></div>;
+        }
+
+        const tradingPair = currentTicker.split('/')
+        const base = window.assetsBySymbol && window.assetsBySymbol[tradingPair[0]]
+        const counter = window.assetsBySymbol && window.assetsBySymbol[tradingPair[1]]
+        const precisions = [base ? base.precision : 5, counter ? counter.precision : 5] || [0,0]
+        var pairBalance = []
+        Object.keys(balance).forEach((currency) => {
+            if (!window.assets[balance[currency].asset]) {
+              return
+            }
+            var item = {}
+            item.currency = window.assets[balance[currency].asset].symbol
+            if (!tradingPair.includes(item.currency)) {
+                return
+            }
+            item.amount = balance[currency].balance ? balance[currency].balance : 0
+            pairBalance.push(item)
+        })
+        const taker_fee = (window.assetsBySymbol && window.assetsBySymbol[tradingPair[trade_side]].options.market_fee_percent/100) || 0
+        const maker_fee = (window.maker_rebate_percent_of_fee && taker_fee - (taker_fee * (window.maker_rebate_percent_of_fee/10000))) || 0
+
+        return (
+            <div className={container + " container-fluid" + (mobile ? " mobile" : "")}>
+                <div className="buy-sell-toggle">
+                    <button id="buy-switch" className={"buy-btn" + (trade_side !== 0 ? " inactive" : "")} onClick={this.switchTradeTo.bind(this, 0)}>BUY <SmallToken name={tradingPair[0]} /></button>
+                    <button id="sell-switch" className={"sell-btn" + (trade_side !== 1 ? " inactive" : "")} onClick={this.switchTradeTo.bind(this, 1)}>SELL <SmallToken name={tradingPair[0]} /></button>
+                </div>
+
+                <div className="transac-actions">
+                    <div className="input-container">
+                        <label>PRICE</label>
+                        <input type="number" className="trade-input qt-number-bold qt-font-small" title=""
+                            name="price"
+                            autoComplete="off"
+                            onFocus={(e) => !mobile && e.target.select()}
+                            min="0"
+                            value={price == "-" ? 0 : price}
+                            onChange={(e) => {
+                              let value = e.target.value
+                              this.setState({
+                              price: value,
+                              total: qty * value
+                            })}} />
+                        <SmallToken name={tradingPair[1]} />
+                    </div>
+                    <div className="input-container">
+                      <label>AMOUNT</label>
+                        <input type="number" className="trade-input qt-number-bold qt-font-small rounded-0-left" title=""
+                            name="amount"
+                            autoComplete="off"
+                            onFocus={(e) => !mobile && e.target.select()}
+                            min="0"
+                            value={Utils.maxPrecision(qty, precisions[0])}
+                            onChange={(e) => {
+                              let value = Utils.maxPrecision(e.target.value, precisions[0])
+                              this.setState({
+                              qty: value,
+                              total: value * price
+                            })}} />
+                        <SmallToken name={tradingPair[0]} />
+                    </div>
+                    <div className="input-container">
+                        <label className="invisible">Percent</label>
+                        <div className="percent-container d-flex justify-content-between text-center">
+                          {dropdown_items.items.map(item => {
+                            return (
+                              <div key={item} className="amount-select cursor-pointer"
+                                onClick={() => balance && this.setPercentAmount(item, tradingPair[trade_side == 0 ? 1 : 0])}
+                              >
+                                {item}
+                              </div>
+                            )
+                          })}
+                        </div>
+                    </div>
+                    <div className="input-container">
+                        <label>TOTAL</label>
+                        <input
+                            title=""
+                            type="number"
+                            autoComplete="off"
+                            onFocus={(e) => !mobile && e.target.select()}
+                            className="trade-input qt-number-bold qt-font-small"
+                            min="0"
+                            value={total < 1/Math.pow(10, precisions[1]) ? total : Utils.maxPrecision(total, precisions[1])}
+                            onChange={(e) => {
+                              let value = Utils.maxPrecision(e.target.value, precisions[1])
+                              this.setState({
+                              qty: value/price,
+                              total: value
+                            })}}
+                            />
+                        <SmallToken name={tradingPair[1]} />
+                    </div>
+                    {total < 1/Math.pow(10, precisions[1]) && !(price <= 0 || qty <= 0) ? 
+                      <div className="text-danger text-right">* total must be more than {1/Math.pow(10, precisions[1])}</div> 
+                      : null
+                    }
+
+                    <div className="fees-container">
+                        Estimate Fees
+                        <table>
+                            <tbody>
+                                <tr>
+                                    <td>Maker</td>
+                                    <td className="text-left text-muted pl-3">{maker_fee}%</td>
+                                    <td className="text-right pr-2">{window.assetsBySymbol 
+                                        && window.assetsBySymbol[tradingPair[trade_side]].options.market_fee_percent != 0 
+                                        && ((trade_side == 0 ? qty : total)*maker_fee/100).toLocaleString(navigator.language, {maximumFractionDigits: precisions[trade_side]}) || 0}</td>
+                                    <td className="text-muted"><SmallToken name={tradingPair[trade_side]} /></td>
+                                </tr>
+                                <tr>
+                                    <td>Taker</td>
+                                    <td className="text-left text-muted pl-3">{taker_fee}%</td>
+                                    <td className="text-right pr-2">{window.assetsBySymbol 
+                                        && window.assetsBySymbol[tradingPair[trade_side]].options.market_fee_percent != 0 
+                                        && ((trade_side == 0 ? qty : total)*taker_fee/100).toLocaleString(navigator.language, {maximumFractionDigits: precisions[trade_side]}) || 0}</td>
+                                    <td className="text-muted"><SmallToken name={tradingPair[trade_side]} /></td>
+                                </tr>
+                                <tr>
+                                    <td colSpan="2">Platform Fees</td>
+                                    <td className="text-right pr-2">{fee.amount}</td>
+                                    <td className="text-muted"><SmallToken name={fee.symbol} /></td>
+                                </tr>
+                            </tbody>
+                        </table>
+                    </div>
+
+                    { private_key ? 
+                      <div>
+                        {trade_side == 1 ?
+                          <button id="sell-action" className="sell-btn" 
+                            disabled={price <= 0 || qty <= 0 || total < 1/Math.pow(10, precisions[1]) || processing}
+                            onClick={this.handleSell.bind(this)}>
+                            {processing ? <Loader /> : 
+                              <span>
+                                PLACE SELL ORDER
+                              </span>
+                            }
+                          </button>
+                          :
+                          <button id="buy-action" className="buy-btn" 
+                            disabled={price <= 0 || qty <= 0 || total < 1/Math.pow(10, precisions[1]) || processing}
+                            onClick={this.handleBuy.bind(this)}>
+                            {processing ? <Loader /> : 
+                              <span>
+                                PLACE BUY ORDER
+                              </span>
+                            }
+                          </button>
+                        }
+                      </div>
+
+                      : 
+                      mobile || publicKey ? 
+                        <button className="connect-btn" 
+                        onClick={mobile && mobile_nav ? () => mobile_nav() 
+                          : 
+                          () => this.props.dispatch({
+                            type: TOGGLE_CONNECT_DIALOG,
+                            data: "connect"
+                          })
+                        }>
+                        CONNECT WALLET TO TRADE
+                      </button> 
+                      : null
+                    }
+
+                    { mobile && publicKey ?
+                      <div className="d-flex justify-content-around flex-wrap qt-font-light qt-font-small text-secondary">
+                        <span className="mx-2">{base.symbol.split("0X")[0]} Balance: {balance[base.id] ? balance[base.id].balance : 0}</span>
+                        <span className="mx-2">{counter.symbol.split("0X")[0]} Balance: {balance[counter.id] ? balance[counter.id].balance : 0}</span>
+                      </div>
+                      : null
+                    }
+                </div>
+            </div>
+        )
+    }
 }
 
 const mapStateToProps = (state) => ({
+    private_key: state.app.private_key,
+    publicKey: state.app.publicKey,
     requestedPrice: 0,
     requestedQty: 0,
-    state: state,
-  	bids: state.app.tradeBook.bids,
-  	asks: state.app.tradeBook.asks,
-		currentPrice: state.app.mostRecentTrade.price,
-		currentTicker:state.app.currentTicker,
-    balance:state.app.balance,
+    bids: state.app.tradeBook.bids,
+    asks: state.app.tradeBook.asks,
+    currentTicker: state.app.currentTicker,
+    currentPrice: state.app.mostRecentTrade,
+    balance: state.app.balance && lodash.keyBy(state.app.balance, "asset"),
     inputBuy: state.app.inputBuy,
     inputBuyAmount: state.app.inputBuyAmount,
     inputSetTime: state.app.setTime,
-    inputSide: state.app.inputSide
-	});
+    inputSide: state.app.inputSide,
+    fee: state.app.fee
+});
 
 export default connect(mapStateToProps)(Trade);
