@@ -8,43 +8,49 @@ import QTTabBar from './ui/tabBar.jsx'
 import QTTableViewSimple from './ui/tableViewSimple.jsx'
 
 const container = css`
-	height: 360px;
-	padding: 18px 20px;
+	height: 100%;
+	padding: 10px;
 	h4 {
 		padding-bottom: 10px;
 		border-bottom: 1px solid #333;
-		margin-bottom: 10px;
+		margin-bottom: 0;
 	}
 	.trading-history-table-container {
-		height: 88%;
+		height: calc(100% - 30px);
 		
 		thead th {
 			position: sticky;
 			position: -webkit-sticky;
 			top: 0;
-			background: #191e21;
+			background: #23282c;
 			z-index: 1;
 		}
 	}
 
 	&.mobile {
-		height: 100%;
-
+		height: auto;
 		thead th {
-			background: #22282c;
+			background: #0A121E;
+		}
+
+		.perc-bar.left {
+			left: 0;
+		}
+
+		.perc-bar.right {
+			right: 0;
 		}
 	}
 `;
 
 class TradingHistory extends Component {
 	setAmount(e) {
-		console.log(e)
 		const set_time = new Date();
 		this.props.dispatch({
 			type: 'SET_AMOUNT',
 			data: {
 				inputBuy: e.price,
-				inputBuyAmount: e.amount.replace(/,/g, ""),
+				inputBuyAmount: e.amount,
 				inputSide: e.color_key,
 				setTime: set_time
 			}
@@ -56,7 +62,7 @@ class TradingHistory extends Component {
 			names: ['TRADING HISTORY','MY TRADES'],
 			selectedTabIndex: 0,
 		}
-
+		
 		return (
 			<div className={container + (this.props.mobile ? " mobile" : "")}>
 				{/* <QTTabBar
@@ -64,12 +70,16 @@ class TradingHistory extends Component {
 					width={122.9}
 					tabs = {tabs}
 				/> */}
-				<h4>TRADING HISTORY</h4>
-				<section className="trading-history-table-container no-scroll-bar">
+				<h4>TRADE HISTORY</h4>
+				<section className="trading-history-table-container no-scroll-bar pushed-margin">
 					
 					<div>
-						<QTTableViewSimple dataSource={this.props.trades.dataSource} columns={this.props.trades.columns}
-											onAction={this.setAmount.bind(this)} />
+						<QTTableViewSimple 
+							max={this.props.trades.max}
+							barDir="left"
+							dataSource={this.props.trades.dataSource} 
+							columns={this.props.trades.columns}
+							onAction={this.setAmount.bind(this)} />
 					</div>
 				</section>
 			</div>
