@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { css } from 'emotion'
 
-const loader = css `
+const loader = css`
     border: 3px solid rgba(255,255,255,0.5);
     border-radius: 50%;
     border-top: 3px solid #fff;
@@ -22,10 +22,59 @@ const loader = css `
     }
 `
 
+const fb_loader = css`
+    .lds-facebook {
+        display: inline-block;
+        position: relative;
+        width: 64px;
+        height: 64px;
+    }
+    .lds-facebook div {
+        display: inline-block;
+        position: absolute;
+        left: 6px;
+        width: 13px;
+        background: rgba(0,0,0,0.5);
+        animation: lds-facebook 1.2s cubic-bezier(0, 0.5, 0.5, 1) infinite;
+    }
+    .lds-facebook div:nth-child(1) {
+        left: 6px;
+        animation-delay: -0.24s;
+    }
+    .lds-facebook div:nth-child(2) {
+        left: 26px;
+        animation-delay: -0.12s;
+    }
+    .lds-facebook div:nth-child(3) {
+        left: 45px;
+        animation-delay: 0;
+    }
+    @keyframes lds-facebook {
+        0% {
+        top: 6px;
+        height: 51px;
+        }
+        50%, 100% {
+        top: 19px;
+        height: 26px;
+        }
+    }
+  
+`
+
 export default class Loader extends Component {
     render() {
-        return (
-            <div className={loader} style={{width: this.props.size, height: this.props.size, margin: this.props.margin}}></div>
-        )
+        if (this.props.type == "box") {
+            return (
+                <div className={fb_loader} style={{width: this.props.size, height: this.props.size, margin: this.props.margin}}>
+                    <div className="lds-facebook"><div></div><div></div><div></div></div>
+                </div>
+            )
+        } else {
+            return (
+                <div className={loader} style={{width: this.props.size, height: this.props.size, margin: this.props.margin}}></div>
+            )
+        }
+        
     }
 }
