@@ -7,13 +7,13 @@ const container = css `
     text-align: left;
     padding: 0 35px 0 20px;
     font-size: 13px;
-    background: url(../public/images/search.png) no-repeat calc(100% - 15px) 50%;
+    background: url(${devicePath("public/images/search.png")}) no-repeat calc(100% - 15px) 50%;
 `
 
 export default class SearchBox extends Component {
     render() {
         return (
-            <input className={container} spellCheck="false"
+            <input className={container} spellCheck="false" autoComplete="off" onClick={(e) => e.stopPropagation()}
 				onChange={this.props.onChange} placeholder={this.props.placeholder} style={this.props.style} />
         )
     }
